+++ NEWS.md
@@ -1,5 +1,9 @@
 # TidierDB.jl updates
-## v.05.2 - 2014 2024-12-03
+
+## v.05.2 - 2024-12-10
+- adds `@relocate`
+
+## v.05.2 - 2024-12-03
 - adds support `_by` support to `@mutate` and `@summarize` for grouping within the macro call.
 - adds support for `n()` in `@mutate`
 - add support for unnesting content to mutate/filter etc via `column[key]`syntax